\name{extract_info_op}
\alias{extract_info_op}
\title{function extract_info_op}
\usage{
extract_info_op(OSM.Data,value)
}
\arguments{
\item{OSM.Data}{an osm object.}
\item{value}{a map feature.}

}
\value{.
}
\description{

}
\author{
Jan-Philipp Kolb \email{Jan-Philipp_Kolb@gmx.net}
}
\examples{


info <- extract_info_op(OSM.Data=schools_MA,value="school")

}
